+++ src/examples/zkapps/reducer/reducer-composite.ts
@@ -7,7 +7,6 @@ import {
   SmartContract,
   Mina,
   AccountUpdate,
-  isReady,
   Bool,
   Struct,
   Reducer,
@@ -15,8 +14,6 @@ import {
 } from 'o1js';
 import assert from 'node:assert/strict';
 import { getProfiler } from '../../utils/profiler.js';
-
-await isReady;
 
 class MaybeIncrement extends Struct({
   isIncrement: Bool,