+++ src/examples/zkapps/reducer/reducer.ts
@@ -83,7 +83,7 @@ if (doProofs) {
   // TODO: if we don't do this, then `analyzeMethods()` will be called during `runUnchecked()` in the tx callback below,
   // which currently fails due to `finalize_is_running` in snarky not resetting internal state, and instead setting is_running unconditionally to false,
   // so we can't nest different snarky circuit runners
-  CounterZkapp.analyzeMethods();
+  await CounterZkapp.analyzeMethods();
 }
 
 console.log('deploy');